% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rank.R
\name{compute_rank}
\alias{compute_rank}
\title{Compute the rank of each value in a vector based on quantiles.}
\usage{
compute_rank(x, n = 10)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{The number of quantiles to use for ranking. Default is 10.}
}
\value{
A numeric vector of the same length as x, with each value replaced by its rank.
}
\description{
This function computes the rank of each value in a vector based on quantiles.
}
\note{
If the length of x is smaller than n, the function will stop and return an error.
}
\examples{
fun.compute_rank(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

}
