% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{create_bal_panel}
\alias{create_bal_panel}
\alias{make_bal_panel}
\title{For a given panel data, create a balanced panel using a set of variables}
\usage{
create_bal_panel(df, id, time, vars)

make_bal_panel(df, id, time, vars)
}
\arguments{
\item{df}{A data frame}

\item{id}{A unique identifier of the panel data, string}

\item{time}{A time variable, string}

\item{vars}{A set of variables to be used to create the panel balanced}
}
\description{
For a given panel data, create a balanced panel using a set of variables
}
\examples{
\donttest{
df <- tibble::tribble(
  ~ id, ~ year, ~ x, ~ y, ~ z,
  1, 2000, 1, 2, NA_real_,
  1, 2001, 1, 2, 3,
  2, 2000, 1, NA_real_, 3,
  2, 2001, 1, 2, 3
)

create_bal_panel(df, "id", "year", c(x))
create_bal_panel(df, "id", "year", c(x, y))
}
}
