% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{create_join}
\alias{create_join}
\title{Create the join (finest common coarsening) of two partitions}
\usage{
create_join(p1, p2)
}
\arguments{
\item{p1}{Partition object created by \code{create_partition}.}

\item{p2}{Partition object created by \code{create_partition}.}
}
\value{
A partition object representing the join of p1 and p2.
}
\description{
This function creates the join of two partitions, which is the finest
partition that is coarser than both input partitions. In other words,
it finds the finest common coarsening of p1 and p2.
}
\details{
The join is computed by finding the finest partition such that both
p1 and p2 are refinements of it. This means finding cutoffs that
appear in both partitions (common boundaries).
}
\examples{
# Example 1
p1 <- create_partition(list(c(0, 1), c(1, 3), c(3, 4)))
p2 <- create_partition(list(c(0, 2), c(2, 4)))
join_p <- create_join(p1, p2)  # Result: [0,4)
is_finer(p1, join_p)  # TRUE
is_finer(p2, join_p)  # TRUE

# Example 2
p3 <- create_partition(list(c(0, 1), c(1, 2), c(2, 3)))
p4 <- create_partition(list(c(0, 1), c(1, 3)))
join_p2 <- create_join(p3, p4)  # Result: [0,1), [1,3)
is_finer(p3, join_p2)  # TRUE
is_finer(p4, join_p2)  # TRUE
}
