% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{create_meet}
\alias{create_meet}
\title{Create the meet (coarsest common refinement) of two partitions}
\usage{
create_meet(p1, p2)
}
\arguments{
\item{p1}{Partition object created by \code{create_partition}.}

\item{p2}{Partition object created by \code{create_partition}.}
}
\value{
A partition object representing the meet of p1 and p2.
}
\description{
This function creates the meet of two partitions, which is the coarsest
partition that is finer than both input partitions. In other words,
it finds the coarsest common refinement of p1 and p2.
}
\details{
The meet is computed by taking all cutoff points from both partitions
and creating the partition that includes all these boundaries. This is
the coarsest partition that is finer than both p1 and p2.
}
\examples{
# Example 1
p1 <- create_partition(list(c(0, 1), c(1, 3), c(3, 4)))
p2 <- create_partition(list(c(0, 2), c(2, 4)))
meet_p <- create_meet(p1, p2)  # Result: [0,1), [1,2), [2,3), [3,4)

# Example 2
p3 <- create_partition(list(c(0, 2), c(2, 4)))
p4 <- create_partition(list(c(0, 1), c(1, 4)))
meet_p2 <- create_meet(p3, p4)  # Result: [0,1), [1,2), [2,4)
}
