% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_panel.R
\name{create_panel}
\alias{create_panel}
\title{create a balance panel}
\usage{
create_panel(N, Time, J = 1)
}
\arguments{
\item{N}{number of firms}

\item{Time}{number of periods}

\item{J}{number of groups (industries)}
}
\description{
create a balance panel
}
