% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{create_partition}
\alias{create_partition}
\title{Create a partition of the real line}
\usage{
create_partition(intervals)
}
\arguments{
\item{intervals}{List of intervals, where each interval is a vector of 2 numbers}
}
\value{
A partition object
}
\description{
Create a partition of the real line
}
\examples{
examples/intervals_example.R
}
