% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{create_partition_from_cutoffs}
\alias{create_partition_from_cutoffs}
\title{Create a partition from cutoff points}
\usage{
create_partition_from_cutoffs(cutoffs)
}
\arguments{
\item{cutoffs}{A numeric vector of cutoff points that define the boundaries
of the intervals. Must be sorted in ascending order.}
}
\value{
A partition object with intervals defined by consecutive cutoffs.
}
\description{
This function creates a partition from a vector of cutoff points by creating
intervals between consecutive cutoffs.
}
\details{
The function creates intervals of the form \code{[cutoffs[i], cutoffs[i+1])} for
i = 1, ..., length(cutoffs)-1. The cutoffs must be sorted in ascending order.
Infinite values are allowed for the first and last cutoffs.
}
\examples{
# Create partition from cutoffs
cutoffs <- c(-Inf, 0, 1, 2, Inf)
p <- create_partition_from_cutoffs(cutoffs)
#  (-Inf, 0), [0, 1), [1, 2), [2, Inf)

# Finite cutoffs
cutoffs2 <- c(0, 1, 2, 3)
p2 <- create_partition_from_cutoffs(cutoffs2)
#  [0, 1), [1, 2), [2, 3)
}
