% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_outlier.r
\name{fun.drop_outlier}
\alias{fun.drop_outlier}
\title{Replaces outliers in a numeric vector with NA.}
\usage{
fun.drop_outlier(x, p = 0.01)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{The percentile cutoff to use for defining outliers. Default is 0.01.}
}
\value{
A numeric vector of the same length as x, with outliers replaced by NA.
}
\description{
This function replaces outliers in the vector x with NA. Outliers are defined as values
that are below the p-th percentile or above the (1-p)-th percentile.
}
\examples{
fun.drop_outlier(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

}
