% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{get_lengths}
\alias{get_lengths}
\title{Get lengths of intervals in a partition}
\usage{
get_lengths(partition)
}
\arguments{
\item{partition}{Partition object created by \code{create_partition}.}
}
\value{
A numeric vector containing the length of each interval.
}
\description{
This function calculates the length of each interval in a partition.
For intervals with infinite bounds, the length is returned as Inf.
}
\examples{
p <- create_partition(list(c(-Inf, 1), c(1, 2), c(2, Inf)))
get_lengths(p)  # c(Inf, 1, Inf)

p2 <- create_partition(list(c(0, 1), c(1, 2), c(2, 3)))
get_lengths(p2)  # c(1, 1, 1)
}
