% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{get_n_intervals}
\alias{get_n_intervals}
\title{Get the number of intervals in a partition}
\usage{
get_n_intervals(partition)
}
\arguments{
\item{partition}{Partition object created by \code{create_partition}.}
}
\value{
A numeric value containing the number of intervals.
}
\description{
This function returns the number of intervals in a partition.
}
\examples{
p <- create_partition(list(c(-Inf, 1), c(1, 2), c(2, Inf)))
get_n_intervals(p) # 3

p2 <- create_partition(list(c(0, 1), c(1, 2), c(2, 3)))
get_n_intervals(p2) # 3
}
