% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_path.R
\name{get_script_path}
\alias{get_script_path}
\title{Get the directory path of the current file}
\usage{
get_script_path()
}
\value{
A character string containing the directory path of the current file
}
\description{
This function returns the directory path of the current R script being executed.
It works both when the script is run from the command line (using Rscript) and
when run interactively in RStudio.
}
\examples{
\dontrun{
# Get the directory of the current file
current_dir <- get_script_path()
}
}
