% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitions.R
\name{is_finer}
\alias{is_finer}
\title{Check if one partition is finer than another}
\usage{
is_finer(p1, p2)
}
\arguments{
\item{p1}{Partition object created by \code{create_partition}.}

\item{p2}{Partition object created by \code{create_partition}.}
}
\value{
Logical value: TRUE if p1 is finer than p2, FALSE otherwise.
}
\description{
This function determines if partition p1 is finer than partition p2.
A partition p1 is finer than p2 if every interval in p1 is contained
within some interval in p2.
}
\details{
For p1 to be finer than p2, every interval [a, b) in p1 must be
completely contained within some interval [c, d) in p2, meaning
c <= a < b <= d.
}
\examples{
p1 <- create_partition(list(c(0, 1), c(1, 2), c(2, 3), c(3, 4)))
p2 <- create_partition(list(c(0, 2), c(2, 4)))
is_finer(p1, p2) # TRUE
is_finer(p2, p1) # FALSE
}
