% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bal_panel.R
\name{make_bal_panel}
\alias{make_bal_panel}
\title{For a given panel data, make a balanced panel using a set of variables}
\usage{
make_bal_panel(df, id, time, vars)
}
\arguments{
\item{df}{A data frame}

\item{id}{A unique identifier of the panel data, string}

\item{time}{A time variable, string}

\item{vars}{A set of variables to be used to make the panel balanced}
}
\description{
For a given panel data, make a balanced panel using a set of variables
}
