% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Generate a custom grid of values with a concentration effect.}
\usage{
make_grid(min_val, max_val, length, concentration = 2, center = min_val)
}
\arguments{
\item{min_val}{A numeric value specifying the minimum value of the grid.}

\item{max_val}{A numeric value specifying the maximum value of the grid.}

\item{length}{An integer specifying the number of grid points to generate.}

\item{concentration}{A numeric value controlling the concentration of grid points.
For \code{center = min_val} or \code{center = max_val}, higher values yield more concentration.
For an arbitrary \code{center}, the sum of the Beta distribution parameters is proportional to \code{concentration}.
Default is 2.}

\item{center}{A numeric value between \code{min_val} and \code{max_val} specifying the value around which
the grid points will be concentrated. Default is \code{min_val}.}
}
\value{
A numeric vector of length \code{n} containing the generated grid values.
}
\description{
This function creates a numeric grid that spans from a specified minimum value to a maximum value,
with a specified number of points. A concentration parameter controls how the points are distributed.
The grid can be generated with a higher concentration of points around:
(i) \code{min_val} (by setting \code{center = min_val}),
(ii) \code{max_val} (by setting \code{center = max_val}), or
(iii) an arbitrary value between \code{min_val} and \code{max_val} (by setting \code{center} to that value).
}
\details{
For \code{center = min_val} or \code{center = max_val}, a power transformation is used.
For an arbitrary \code{center}, the Beta distribution quantile function is used to achieve the desired concentration.
}
\examples{
# Set up parameters
min_val <- 0
max_val <- 100
length <- 100
concentration <- 5

# Generate grids with different centers
grid_min <- make_grid(min_val, max_val, length, concentration, center = min_val)
grid_mid <- make_grid(min_val, max_val, length, concentration, center = 70)
grid_max <- make_grid(min_val, max_val, length, concentration, center = max_val)

# Plotting the grids in separate panels
par(mfrow = c(3, 1), mar = c(4, 4, 2, 1))

# Grid concentrated near the minimum value
plot(grid_min, rep(1, length), type = "b", xlim = c(min_val, max_val), 
     xlab = "Grid Values", ylab = "", main = "Concentration around min_val")
grid()

# Grid concentrated around an arbitrary value (70)
plot(grid_mid, rep(1, length), type = "b", xlim = c(min_val, max_val), 
     xlab = "Grid Values", ylab = "", main = "Concentration around 70")
grid()

# Grid concentrated near the maximum value
plot(grid_max, rep(1, length), type = "b", xlim = c(min_val, max_val), 
     xlab = "Grid Values", ylab = "", main = "Concentration around max_val")
grid()

}
