\name{matsuno-package}
\alias{matsuno-package}
\alias{matsuno}
\docType{package}
\title{
\packageTitle{matsuno}
}
\description{
\packageDescription{matsuno}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matsuno}
\packageIndices{matsuno}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{matsuno}

Maintainer: \packageMaintainer{matsuno}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
