% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_discontinuity.R
\name{preprocess_discontinuity}
\alias{preprocess_discontinuity}
\title{Preprocess data to handle discontinuities}
\usage{
preprocess_discontinuity(data, ..., threshold = 0.01)
}
\arguments{
\item{data}{A data frame containing the variables to process.}

\item{...}{Unquoted variable names to process for discontinuities.}

\item{threshold}{A numeric value specifying the threshold for detecting discontinuities.}
}
\value{
A data frame with \code{NA} introduced in the specified variables where discontinuities are detected.
}
\description{
This function introduces \code{NA} in specified variables (\code{...}) of a data frame
wherever the absolute difference between consecutive values exceeds a threshold.
}
\examples{
\dontrun{
x <- seq(-10, 10, length.out = 500)
y <- ifelse(x < 0, sin(x), cos(x)) # Artificial discontinuity at x = 0
z <- ifelse(x < 0, cos(x), sin(x)) # Another artificial discontinuity
data <- data.frame(x = x, y = y, z = z)

data <- preprocess_discontinuity(data, vars = c("y", "z"), threshold = 0.5)
head(data)
}
}
