% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptruncnorm_cpp}
\alias{ptruncnorm_cpp}
\title{Calculate truncated normal cumulative distribution function}
\usage{
ptruncnorm_cpp(x, mean = 0, sd = 1, a, b)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mean}{Mean of the normal distribution before truncation (default: 0).}

\item{sd}{Standard deviation of the normal distribution before truncation (default: 1).}

\item{a}{Lower truncation point.}

\item{b}{Upper truncation point.}
}
\value{
Vector of probabilities corresponding to the CDF evaluated at each value in x.
}
\description{
Computes the cumulative distribution function of a truncated normal
distribution in a numerically stable way.
}
\details{
Memo: the defualt values of \code{a} and \code{b} do not work. The only solution seems to create a wrapper of this using R.
Basically, we cannot set the default values to be Inf. Now, I specify R_NegInf and R_PosInf, but these are not parsed.
If I do not specify default values, the function is not compiled. So, this is a hacky way to make it work.
}
