% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme_original.R
\name{theme_matsuno}
\alias{theme_matsuno}
\title{ggplot2 theme original}
\usage{
theme_matsuno(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\description{
ggplot2 theme original
}
