% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkgs.R, R/utilities.R
\name{update_pkgs}
\alias{update_pkgs}
\title{Update all packages}
\usage{
update_pkgs(exclude_pkg = "arrow")

update_pkgs(exclude_pkg = "arrow")
}
\arguments{
\item{exclude_pkg}{Character vector of package names to exclude from updates}
}
\description{
This function checks for package updates, lists them, and asks for user confirmation.

This function checks for package updates, lists them, and asks for user confirmation.
}
