% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{get_interval}
\alias{get_interval}
\title{Get which interval a number falls into}
\usage{
get_interval(x, partition)
}
\arguments{
\item{x}{Number to check}

\item{partition}{Partition object}
}
\value{
A list with the index of the interval and the interval itself, or NA if not found
}
\description{
Get which interval a number falls into
}
\examples{
p <- create_partition(list(c(-Inf, 1), c(1, 2), c(2, Inf)))
get_interval(0.5, p)  # list(index = 1, interval = c(-Inf, 1))
get_interval(1.5, p)  # list(index = 2, interval = c(1, 2))
get_interval(2.5, p)  # list(index = 3, interval = c(2, Inf))
}
