% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_original.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate two strings
Can take an expression in argument: (1+2) %+% " unit" returns "3 unit"}
\usage{
s1 \%+\% s2
}
\description{
Concatenate two strings
Can take an expression in argument: (1+2) %+% " unit" returns "3 unit"
}
