% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bal_panel.R
\name{make_bal_panel}
\alias{make_bal_panel}
\title{For a given panel data, make a balanced panel using a set of variables}
\usage{
make_bal_panel(df, id, time, vars)
}
\arguments{
\item{df}{A data frame}

\item{vars}{A set of variables to be used to make the panel balanced}

\item{key}{A unique identifier of the panel data, string}
}
\description{
For a given panel data, make a balanced panel using a set of variables
}
\examples{
df <- tibble::tribble(
  ~ id, ~ year, ~ x, ~ y, ~ z,
  1, 2000, 1, 2, NA_real_,
  1, 2001, 1, 2, 3,
  2, 2000, 1, NA_real_, 3,
  2, 2001, 1, 2, 3
)

make_bal_panel(df, "id", "year", c(x))
make_bal_panel(df, "id", "year", c(x, y))
}
