% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{make_grid}
\alias{make_grid}
\alias{create_grid}
\title{Deprecated Alias for create_grid}
\usage{
make_grid(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{create_grid}}.}
}
\description{
This function is an alias for \code{\link{create_grid}} and will be removed in a future version.
}
\seealso{
\code{\link{create_grid}}
}
